Corrected sentence locator regex for "so" and "there is" detection.
Sentences for the "so" and "there is" starting detector may
no longer start with whitespace.

The leading whitespace was causing an unexpected index that was
incremented by the number of leading newlines in a file. This placed
the error in an odd place, causing a lot of problems down the chain.

This patch ignores the concept of a "leading newline" and includes
some tests for making sure this works.